+++ demos/transformer/src/main/java/androidx/media3/demo/transformer/TransformerActivity.java
@@ -88,7 +88,8 @@ import androidx.media3.transformer.Effects;
 import androidx.media3.transformer.ExperimentalAnalyzerModeFactory;
 import androidx.media3.transformer.ExportException;
 import androidx.media3.transformer.ExportResult;
-import androidx.media3.transformer.InAppMuxer;
+import androidx.media3.transformer.InAppFragmentedMp4Muxer;
+import androidx.media3.transformer.InAppMp4Muxer;
 import androidx.media3.transformer.JsonUtil;
 import androidx.media3.transformer.ProgressHolder;
 import androidx.media3.transformer.Transformer;
@@ -302,12 +303,12 @@ public final class TransformerActivity extends AppCompatActivity {
         transformerBuilder.setMaxDelayBetweenMuxerSamplesMs(C.TIME_UNSET);
       }
 
-      if (bundle.getBoolean(ConfigurationActivity.USE_MEDIA3_MUXER)) {
-        transformerBuilder.setMuxerFactory(
-            new InAppMuxer.Factory.Builder()
-                .setOutputFragmentedMp4(
-                    bundle.getBoolean(ConfigurationActivity.PRODUCE_FRAGMENTED_MP4))
-                .build());
+      if (bundle.getBoolean(ConfigurationActivity.USE_MEDIA3_MP4_MUXER)) {
+        transformerBuilder.setMuxerFactory(new InAppMp4Muxer.Factory());
+      }
+
+      if (bundle.getBoolean(ConfigurationActivity.USE_MEDIA3_FRAGMENTED_MP4_MUXER)) {
+        transformerBuilder.setMuxerFactory(new InAppFragmentedMp4Muxer.Factory());
       }
 
       if (bundle.getBoolean(ConfigurationActivity.ENABLE_DEBUG_PREVIEW)) {