+++ libraries/transformer/src/main/java/androidx/media3/transformer/InAppFragmentedMp4Muxer.java
@@ -15,11 +15,8 @@
  */
 package androidx.media3.transformer;
 
-import static androidx.media3.muxer.Mp4Muxer.LAST_SAMPLE_DURATION_BEHAVIOR_SET_FROM_END_OF_STREAM_BUFFER_OR_DUPLICATE_PREVIOUS;
-
 import android.media.MediaCodec;
 import android.media.MediaCodec.BufferInfo;
-import androidx.annotation.Nullable;
 import androidx.media3.common.C;
 import androidx.media3.common.Format;
 import androidx.media3.common.Metadata;
@@ -28,7 +25,6 @@ import androidx.media3.common.util.Log;
 import androidx.media3.common.util.UnstableApi;
 import androidx.media3.container.Mp4OrientationData;
 import androidx.media3.muxer.FragmentedMp4Muxer;
-import androidx.media3.muxer.Mp4Muxer;
 import androidx.media3.muxer.Muxer;
 import androidx.media3.muxer.MuxerException;
 import androidx.media3.muxer.MuxerUtil;
@@ -37,80 +33,15 @@ import com.google.errorprone.annotations.CanIgnoreReturnValue;
 import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.nio.ByteBuffer;
-import java.util.LinkedHashSet;
 import java.util.Locale;
-import java.util.Set;
 
-/** {@link Muxer} implementation that uses an {@link Mp4Muxer} or {@link FragmentedMp4Muxer}. */
+/** {@link Muxer} implementation that uses a {@link FragmentedMp4Muxer}. */
+// TODO: b/372417042 - Add E2E tests for producing fragmented MP4 output.
 @UnstableApi
-public final class InAppMuxer implements Muxer {
-
-  /** Provides {@linkplain Metadata.Entry metadata} to add in the output MP4 file. */
-  public interface MetadataProvider {
-
-    /**
-     * Updates the list of {@linkplain Metadata.Entry metadata entries}.
-     *
-     * <p>A {@link Metadata.Entry} can be added or removed. To modify an existing {@link
-     * Metadata.Entry}, first remove it and then add a new one.
-     *
-     * <p>For the list of supported metadata refer to {@link
-     * Mp4Muxer#addMetadataEntry(Metadata.Entry)}.
-     */
-    void updateMetadataEntries(Set<Metadata.Entry> metadataEntries);
-  }
-
-  /** {@link Muxer.Factory} for {@link InAppMuxer}. */
+public final class InAppFragmentedMp4Muxer implements Muxer {
+  /** {@link Muxer.Factory} for {@link InAppFragmentedMp4Muxer}. */
   public static final class Factory implements Muxer.Factory {
-
-    /** A builder for {@link Factory} instances. */
-    public static final class Builder {
-      @Nullable private MetadataProvider metadataProvider;
-      private boolean outputFragmentedMp4;
-      private long fragmentDurationMs;
-
-      /** Creates a {@link Builder} instance with default values. */
-      public Builder() {
-        fragmentDurationMs = C.TIME_UNSET;
-      }
-
-      /**
-       * Sets an implementation of {@link MetadataProvider}.
-       *
-       * <p>The default value is {@code null}.
-       *
-       * <p>If the value is not set then the {@linkplain Metadata.Entry metadata} from the input
-       * file is set as it is in the output file.
-       */
-      @CanIgnoreReturnValue
-      public Builder setMetadataProvider(MetadataProvider metadataProvider) {
-        this.metadataProvider = metadataProvider;
-        return this;
-      }
-
-      /** Sets whether to output a fragmented MP4. */
-      @CanIgnoreReturnValue
-      public Builder setOutputFragmentedMp4(boolean outputFragmentedMp4) {
-        this.outputFragmentedMp4 = outputFragmentedMp4;
-        return this;
-      }
-
-      /**
-       * Sets the fragment duration (in milliseconds) if the output file is {@link
-       * #setOutputFragmentedMp4(boolean) fragmented}.
-       */
-      @CanIgnoreReturnValue
-      public Builder setFragmentDurationMs(long fragmentDurationMs) {
-        this.fragmentDurationMs = fragmentDurationMs;
-        return this;
-      }
-
-      /** Builds a {@link Factory} instance. */
-      public Factory build() {
-        return new Factory(metadataProvider, outputFragmentedMp4, fragmentDurationMs);
-      }
-    }
-
+    // TODO: b/372417042 - Move these lists to FragmentedMp4Muxer.
     /** A list of supported video sample MIME types. */
     private static final ImmutableList<String> SUPPORTED_VIDEO_SAMPLE_MIME_TYPES =
         ImmutableList.of(
@@ -129,18 +60,21 @@ public final class InAppMuxer implements Muxer {
             MimeTypes.AUDIO_OPUS,
             MimeTypes.AUDIO_VORBIS);
 
-    @Nullable private final MetadataProvider metadataProvider;
-    private final boolean outputFragmentedMp4;
     private final long fragmentDurationMs;
 
     private long videoDurationUs;
 
-    private Factory(
-        @Nullable MetadataProvider metadataProvider,
-        boolean outputFragmentedMp4,
-        long fragmentDurationMs) {
-      this.metadataProvider = metadataProvider;
-      this.outputFragmentedMp4 = outputFragmentedMp4;
+    /** Creates an instance with default values. */
+    public Factory() {
+      this(C.TIME_UNSET);
+    }
+
+    /**
+     * Creates an instance.
+     *
+     * @param fragmentDurationMs The fragment duration (in milliseconds).
+     */
+    public Factory(long fragmentDurationMs) {
       this.fragmentDurationMs = fragmentDurationMs;
       videoDurationUs = C.TIME_UNSET;
     }
@@ -167,7 +101,7 @@ public final class InAppMuxer implements Muxer {
     }
 
     @Override
-    public InAppMuxer create(String path) throws MuxerException {
+    public InAppFragmentedMp4Muxer create(String path) throws MuxerException {
       FileOutputStream outputStream;
       try {
         outputStream = new FileOutputStream(path);
@@ -175,23 +109,13 @@ public final class InAppMuxer implements Muxer {
         throw new MuxerException("Error creating file output stream", e);
       }
 
-      Muxer muxer = null;
-      if (outputFragmentedMp4) {
-        FragmentedMp4Muxer.Builder builder = new FragmentedMp4Muxer.Builder(outputStream);
-        if (fragmentDurationMs != C.TIME_UNSET) {
-          builder.setFragmentDurationMs(fragmentDurationMs);
-        }
-        muxer = builder.build();
-      } else {
-        Mp4Muxer.Builder builder = new Mp4Muxer.Builder(outputStream);
-        if (videoDurationUs != C.TIME_UNSET) {
-          builder.setLastSampleDurationBehavior(
-              LAST_SAMPLE_DURATION_BEHAVIOR_SET_FROM_END_OF_STREAM_BUFFER_OR_DUPLICATE_PREVIOUS);
-        }
-        muxer = builder.build();
+      FragmentedMp4Muxer.Builder builder = new FragmentedMp4Muxer.Builder(outputStream);
+      if (fragmentDurationMs != C.TIME_UNSET) {
+        builder.setFragmentDurationMs(fragmentDurationMs);
       }
+      FragmentedMp4Muxer muxer = builder.build();
 
-      return new InAppMuxer(muxer, metadataProvider, videoDurationUs);
+      return new InAppFragmentedMp4Muxer(muxer, videoDurationUs);
     }
 
     @Override
@@ -205,27 +129,22 @@ public final class InAppMuxer implements Muxer {
     }
   }
 
-  private static final String TAG = "InAppMuxer";
+  private static final String TAG = "InAppFragmentedMp4Muxer";
   private static final int TRACK_ID_UNSET = -1;
 
-  private final Muxer muxer;
-  @Nullable private final MetadataProvider metadataProvider;
+  private final FragmentedMp4Muxer muxer;
   private final long videoDurationUs;
-  private final Set<Metadata.Entry> metadataEntries;
 
   private int videoTrackId;
 
-  private InAppMuxer(
-      Muxer muxer, @Nullable MetadataProvider metadataProvider, long videoDurationUs) {
+  private InAppFragmentedMp4Muxer(FragmentedMp4Muxer muxer, long videoDurationUs) {
     this.muxer = muxer;
-    this.metadataProvider = metadataProvider;
     this.videoDurationUs = videoDurationUs;
-    metadataEntries = new LinkedHashSet<>();
     videoTrackId = TRACK_ID_UNSET;
   }
 
   @Override
-  public int addTrack(Format format) throws MuxerException {
+  public int addTrack(Format format) {
     int trackId = muxer.addTrack(format);
     if (MimeTypes.isVideo(format.sampleMimeType)) {
       muxer.addMetadataEntry(new Mp4OrientationData(format.rotationDegrees));
@@ -255,7 +174,7 @@ public final class InAppMuxer implements Muxer {
   @Override
   public void addMetadataEntry(Metadata.Entry metadataEntry) {
     if (MuxerUtil.isMetadataSupported(metadataEntry)) {
-      metadataEntries.add(metadataEntry);
+      muxer.addMetadataEntry(metadataEntry);
     }
   }
 
@@ -270,20 +189,6 @@ public final class InAppMuxer implements Muxer {
           MediaCodec.BUFFER_FLAG_END_OF_STREAM);
       writeSampleData(videoTrackId, ByteBuffer.allocateDirect(0), bufferInfo);
     }
-    writeMetadata();
     muxer.close();
-  }
-
-  private void writeMetadata() {
-    if (metadataProvider != null) {
-      Set<Metadata.Entry> metadataEntriesCopy = new LinkedHashSet<>(metadataEntries);
-      metadataProvider.updateMetadataEntries(metadataEntriesCopy);
-      metadataEntries.clear();
-      metadataEntries.addAll(metadataEntriesCopy);
-    }
-
-    for (Metadata.Entry entry : metadataEntries) {
-      muxer.addMetadataEntry(entry);
-    }
   }
 }